requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-filter==24.2
python-dotenv==1.1.1
# pytz is deprecated but you can keep it if your project needs it:
pytz
reportlab